Lab_2/requirements.txt
numpy
matplotlib
mplcursors
geopy
pyproj
geographiclib
pandas
python-docx